extract api call
